+++ apis/workloads/v1alpha1/integration/cfpackage_webhook_integration_test.go
@@ -1,4 +1,4 @@
-package v1alpha1_test
+package integration_test
 
 import (
 	"context"
