+++ controllers/workloads/suite_test.go
@@ -1,0 +1,13 @@
+package workloads_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestWorkloadsControllers(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Workloads Controllers Unit Test Suite")
+}
